PWN-6355 - Add sell logic on sell payload screen

+++ app/src/main/java/org/p2p/wallet/moonpay/clientsideapi/MoonpayClientSideApi.kt
@@ -8,6 +8,7 @@ import org.p2p.wallet.moonpay.clientsideapi.response.MoonpayTokenCurrencyRespons
 import retrofit2.http.GET
 import retrofit2.http.Path
 import retrofit2.http.Query
+import java.math.BigDecimal
 
 interface MoonpayClientSideApi {
 
@@ -45,6 +46,6 @@ interface MoonpayClientSideApi {
         @Path("token") tokenSymbol: String,
         @Query("apiKey") apiKey: String,
         @Query("quoteCurrencyCode") fiatName: String,
-        @Query("baseCurrencyAmount") tokenAmount: Double
+        @Query("baseCurrencyAmount") tokenAmount: BigDecimal
     ): MoonpaySellQuoteResponse
 }
