+++ app/src/main/java/org/p2p/wallet/moonpay/repository/sell/MoonpaySellRepositoryMapper.kt
@@ -47,12 +47,12 @@ class MoonpaySellRepositoryMapper {
                 tokenName = currencyName,
                 currencyId = currencyId,
                 amounts = MoonpayCurrencyAmounts(
-                    minAmount = minAmount,
-                    maxAmount = maxAmount,
-                    minBuyAmount = minBuyAmount,
-                    maxBuyAmount = maxBuyAmount,
-                    minSellAmount = minSellAmount,
-                    maxSellAmount = maxSellAmount
+                    minAmount = minAmount.toBigDecimal(),
+                    maxAmount = maxAmount.toBigDecimal(),
+                    minBuyAmount = minBuyAmount.toBigDecimal(),
+                    maxBuyAmount = maxBuyAmount.toBigDecimal(),
+                    minSellAmount = minSellAmount.toBigDecimal(),
+                    maxSellAmount = maxSellAmount.toBigDecimal()
                 )
             )
         }
@@ -62,12 +62,12 @@ class MoonpaySellRepositoryMapper {
                 fiatName = currencyName,
                 currencyId = currencyId,
                 amounts = MoonpayCurrencyAmounts(
-                    minAmount = minAmount,
-                    maxAmount = maxAmount,
-                    minBuyAmount = minBuyAmount,
-                    maxBuyAmount = maxBuyAmount,
-                    minSellAmount = minSellAmount,
-                    maxSellAmount = maxSellAmount
+                    minAmount = minAmount.toBigDecimal(),
+                    maxAmount = maxAmount.toBigDecimal(),
+                    minBuyAmount = minBuyAmount.toBigDecimal(),
+                    maxBuyAmount = maxBuyAmount.toBigDecimal(),
+                    minSellAmount = minSellAmount.toBigDecimal(),
+                    maxSellAmount = maxSellAmount.toBigDecimal()
                 )
             )
         }
